+++ CHANGELOG.md
@@ -5,6 +5,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.5.2 - Unreleased
 
+### Added
+
+* `util_seaborn` with helpers and workarounds
+
+### Fixed
+* Issue in adjust gui
+
+
 
 ## Version 0.5.1 - Released 2024-08-28