+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.5.3 - Unreleased
 
+### Fix
+
+* Fixed `ffmpeg_animate_frames` to handle paths and ndarrays
+
 
 ## Version 0.5.2 - Released 2024-09-09